Moved method file_sha256sum closer to its use in namespace app.schema
