+++ source/app/util.py
@@ -21,7 +21,6 @@
 # TODO should probably dispatch the methods provided in this file in the different namespaces
 import base64
 import datetime
-import hashlib
 import logging as log
 import shutil
 import weakref
@@ -95,20 +94,6 @@ def add_obj_history_entry(obj, action, commit=False):
         db.session.commit()
 
     return obj
-
-
-def file_sha256sum(file_path):
-
-    if not Path(file_path).is_file():
-        return None
-
-    sha256_hash = hashlib.sha256()
-    with open(file_path, "rb") as f:
-        # Read and update hash string value in blocks of 4K
-        for byte_block in iter(lambda: f.read(4096), b""):
-            sha256_hash.update(byte_block)
-
-        return sha256_hash.hexdigest().upper()
 
 
 def hmac_sign(data):
